+++ app/upgrades/v16/constants.go
@@ -11,6 +11,12 @@ import (
 // UpgradeName defines the on-chain upgrade name for the Osmosis v16 upgrade.
 const UpgradeName = "v16"
 
+// new token factory parameters
+//
+// at the current gas price of 0.0025uosmo, this corresponds to 0.1 OSMO per
+// denom creation.
+const NewDenomCreationGasConsume uint64 = 40_000_000
+
 var Upgrade = upgrades.Upgrade{
 	UpgradeName:          UpgradeName,
 	CreateUpgradeHandler: CreateUpgradeHandler,