+++ x/poolmanager/types/msgs_test.go
@@ -443,7 +443,6 @@ func TestMsgSplitRouteSwapExactAmountIn(t *testing.T) {
 	for name, tc := range tests {
 		tc := tc
 		t.Run(name, func(t *testing.T) {
-
 			err := tc.msg.ValidateBasic()
 
 			if tc.expectError {
@@ -558,7 +557,6 @@ func TestMsgSplitRouteSwapExactAmountOut(t *testing.T) {
 	for name, tc := range tests {
 		tc := tc
 		t.Run(name, func(t *testing.T) {
-
 			err := tc.msg.ValidateBasic()
 
 			if tc.expectError {